Persistence: [CQL] move all data context operations into CqlStorage class
