+++ src/Abc.Zebus.Persistence.CQL/Storage/CqlStorage.cs
@@ -11,11 +11,12 @@ using Abc.Zebus.Persistence.Storage;
 using Abc.Zebus.Persistence.Util;
 using Abc.Zebus.Util;
 using Cassandra;
+using Cassandra.Data.Linq;
 using log4net;
 
 namespace Abc.Zebus.Persistence.CQL.Storage
 {
-    public class CqlStorage : IStorage, IDisposable
+    public class CqlStorage : ICqlStorage, IDisposable
     {
         private static readonly ILog _log = LogManager.GetLogger(typeof(CqlStorage));
         private static readonly DateTime _unixOrigin = new DateTime(1970, 1, 1, 0, 0, 0, 0);
@@ -131,6 +132,48 @@ namespace Abc.Zebus.Persistence.CQL.Storage
         public void Dispose()
         {
             _parallelPersistor.Dispose();
+        }
+
+        public long GetOldestNonAckedMessageTimestamp(PeerState peer)
+        {
+            var peerId = peer.PeerId.ToString();
+            var lastAckedMessageTimestamp = 0L;
+
+            foreach (var currentBucketId in BucketIdHelper.GetBucketsCollection(peer.OldestNonAckedMessageTimestampInTicks))
+            {
+                var messagesInBucket = _dataContext.PersistentMessages
+                                                   .Where(x => x.PeerId == peerId
+                                                               && x.BucketId == currentBucketId
+                                                               && x.UniqueTimestampInTicks >= peer.OldestNonAckedMessageTimestampInTicks)
+                                                   .OrderBy(x => x.UniqueTimestampInTicks)
+                                                   .Select(x => new { x.IsAcked, x.UniqueTimestampInTicks })
+                                                   .Execute();
+
+                foreach (var message in messagesInBucket)
+                {
+                    if (!message.IsAcked)
+                        return message.UniqueTimestampInTicks;
+
+                    lastAckedMessageTimestamp = message.UniqueTimestampInTicks;
+                    
+                }
+            }
+
+            return lastAckedMessageTimestamp == 0 ? SystemDateTime.UtcNow.Ticks : lastAckedMessageTimestamp + 1;
+        }
+
+        public void CleanBuckets(PeerId peerId, long previousOldestMessageTimestamp, long newOldestMessageTimestamp)
+        {
+            var firstBucketToDelete = BucketIdHelper.GetBucketId(previousOldestMessageTimestamp);
+            var lastBucketToDelete = BucketIdHelper.GetPreviousBucketId(newOldestMessageTimestamp);
+            if (firstBucketToDelete == lastBucketToDelete)
+                return;
+
+            var bucketsToDelete = BucketIdHelper.GetBucketsCollection(firstBucketToDelete, lastBucketToDelete).ToArray();
+            _dataContext.PersistentMessages
+                        .Where(x => x.PeerId == peerId.ToString() && bucketsToDelete.Contains(x.BucketId))
+                        .Delete()
+                        .ExecuteAsync();
         }
 
         private static long ToUnixMicroSeconds(DateTime timestamp)
