+++ src/Abc.Zebus.Persistence.CQL/Storage/ICqlStorage.cs
@@ -1,0 +1,10 @@
+﻿using Abc.Zebus.Persistence.Storage;
+
+namespace Abc.Zebus.Persistence.CQL.Storage
+{
+    public interface ICqlStorage : IStorage
+    {
+        long GetOldestNonAckedMessageTimestamp(PeerState peer);
+        void CleanBuckets(PeerId peerId, long previousOldestMessageTimestamp, long newOldestMessageTimestamp);
+    }
+}
